major: moving Case.java and Query.java classes to the package no.ntnu.mycbr.rest.controller.helper
