+++ src/main/java/no/ntnu/mycbr/rest/controller/helper/Case.java
@@ -1,9 +1,10 @@
-package no.ntnu.mycbr.rest;
+package no.ntnu.mycbr.rest.controller.helper;
 
 import no.ntnu.mycbr.core.Project;
 import no.ntnu.mycbr.core.casebase.Instance;
 import no.ntnu.mycbr.CBREngine;
 import no.ntnu.mycbr.core.model.Concept;
+import no.ntnu.mycbr.rest.App;
 
 import java.util.LinkedHashMap;
